Decoupled user service into service and utils.
