+++ zimmerman/main/service/user/__init__.py
@@ -1,0 +1,1 @@
+# Entry point for the service package
